created custom hook useFavourites to update firebase state on add/remove
